Drop AT API
We are moving to ParaSpell instead
