+++ app/package.json
@@ -31,7 +31,6 @@
     "@sentry/profiling-node": "^8.32.0",
     "@snowbridge/api": "^0.1.19",
     "@snowbridge/contract-types": "^0.1.19",
-    "@substrate/asset-transfer-api": "^0.4.4",
     "@talisman-connect/ui": "^1.1.1",
     "@talisman-connect/wallets": "^1.1.3",
     "@talismn/connect-components": "^1.1.8",
