+++ app/src/__tests__/Transfer.test.tsx
@@ -2,9 +2,9 @@ import { Direction, resolveDirection } from '@/services/transfer'
 import '@testing-library/jest-dom'
 import { Mainnet, mainnetRegistry, Testnet } from '../config/registry'
 import { convertAmount, safeConvertAmount, toHuman } from '@/utils/transfer'
-import { getDestChainId } from '@/hooks/useAssetTransferApi'
 import { Enum } from 'polkadot-api'
 import { getTokenPrice } from '@/services/balance'
+import { getDestChainId } from '@/models/chain'
 
 describe('Transfer', () => {
   it('direction ToEthereum', () => {
@@ -56,13 +56,6 @@ describe('Transfer', () => {
 
   it('convert amount to string', () => {
     expect(convertAmount(0.3, Mainnet.WETH).toString()).toBe('300000000000000000')
-  })
-
-  it('looks up local asset id', () => {
-    expect(mainnetRegistry.localAssetId.get(Mainnet.Bifrost.uid)?.get(Mainnet.WETH.id)).toEqual(
-      Enum('Token2', 13),
-    )
-    //{"type": "Token2", "value": 13})
   })
 
   it('fetches the token price', () => {
