+++ app/src/models/chain.ts
@@ -29,3 +29,13 @@ export enum Network {
 }
 
 export type AddressType = 'evm' | 'ss58'
+
+export function getDestChainId(destChain: Chain): string {
+  switch (destChain.network) {
+    case Network.Ethereum: {
+      return `{"parents":"2","interior":{"X1":{"GlobalConsensus":{"Ethereum":{"chainId":"${destChain.chainId}"}}}}}`
+    }
+    default:
+      return destChain.chainId.toString()
+  }
+}
